Added some ores and their generation. Retextured ores and added some functionality to cursed wool so people can break it in creative without dieing or without explosion .

+++ src/main/java/net/stilt/cursedmobsandpvp/event/CursedWoolAttackEventHandler.java
@@ -7,7 +7,6 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
-import net.minecraft.world.GameMode;
 import net.minecraft.world.World;
 import net.minecraft.world.World.ExplosionSourceType;
 import net.stilt.cursedmobsandpvp.block.ModBlocks;
@@ -17,15 +16,14 @@ public class CursedWoolAttackEventHandler implements AttackBlockCallback{
     @Override
     public ActionResult interact(PlayerEntity player, World world, Hand hand, BlockPos pos, Direction direction) {
         BlockState state = world.getBlockState(pos);
-        GameMode gameMode = GameMode.DEFAULT;
             /* Manual spectator check is necessary because AttackBlockCallbacks
                fire before the spectator check */
-            if (!player.isSpectator() && state.getBlock() == ModBlocks.Cursed_Wool) {
+            if (!player.isSpectator() && state.getBlock() == ModBlocks.Cursed_Wool && !player.getAbilities().creativeMode) {
                 world.createExplosion(null, pos.getX(), pos.getY(), pos.getZ(), 100.0f, ExplosionSourceType.NONE);
                 world.createExplosion(null, pos.getX(), pos.getY() + 1, pos.getZ(), 100.0f, ExplosionSourceType.NONE);
                 world.createExplosion(null, pos.getX(), pos.getY() - 1, pos.getZ(), 100.0f, ExplosionSourceType.NONE);
 
-                if (player.getHealth() != 0f && !player.getAbilities().creativeMode){
+                if (player.getHealth() != 0f){
                     player.kill();
                 }
